"Add Cupertino import for ShipsView"
Explanation:
This code diff adds an import statement for the Cupertino package in the ShipsView file. However, it does not actually use any Cupertino widgets in the code. The diff also changes the Scaffold widget to not use SafeArea, but this might not be related to the Cupertino import. Overall, the diff is adding a new import, but it's unclear if it's being used in the current implementation.